userController setup done...

+++ src/Utils/Cloudinary.js
@@ -4,11 +4,12 @@ import fs from "fs"
 
 cloudinary.config({ 
   cloud_name: process.env.CLOUDINARY_CLOUD_NAME, 
-  api_key: process.env.CLOUDINARY_API_KEY, 
-  api_secret: process.env.CLOUDINARY_API_SECRET 
+  api_key: process.env.CLOUDINARY_CLOUD_API_KEY, 
+  api_secret: process.env.CLOUDINARY_CLOUD_API_SECRET 
 });
 
 const uploadOnCloudinary = async (localFilePath) => {
+    
     try {
         if (!localFilePath) return null
         //upload the file on cloudinary
@@ -16,12 +17,17 @@ const uploadOnCloudinary = async (localFilePath) => {
             resource_type: "auto"
         })
         // file has been uploaded successfull
+
         //console.log("file is uploaded on cloudinary ", response.url);
         fs.unlinkSync(localFilePath)
         return response;
 
     } catch (error) {
-        fs.unlinkSync(localFilePath) // remove the locally saved temporary file as the upload operation got failed
+        console.error('failed to upload File on cloudinary !!')
+        if(existsSync(localFilePath)){
+            fs.unlinkSync(localFilePath)
+        }
+
         return null;
     }
 }
